requirements.txt
numpy
scipy
tensorflow-gpu==1.10
Pillow
matplotlib
tqdm
cvxopt
